비트 패딩, 바이트 패딩 정리

+++ Source/Core/Object/Component.cpp
@@ -52,7 +52,7 @@ namespace sw
 
 	void Component::setTickGroup( TickGroup group )
 	{
-		_tickGroup = group;
+		_tickGroup = static_cast<uint8>( group );
 		if ( _owner != nullptr )
 		{
 			_owner->markTickOrderDirty();
